feat: support multiline string literals

+++ src/source_formatter.rs
@@ -216,7 +216,7 @@ fn find_last_error_node<'tree>(node: &Node<'tree>) -> Option<Node<'tree>> {
     let mut last_error_node = Some(*node);
 
     for i in 0..node.child_count() {
-        if let Some(child) = node.child(i) {
+        if let Some(child) = node.child(i as u32) {
             if child.has_error() {
                 last_error_node = find_last_error_node(&child);
             }
